Place tests into separate modules
